+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change subnet selection condition for transit gateway attachment.
+
 ## [1.4.0] - 2023-01-19
 
 ### Changed